// src/app/models/material.ts
export class Material {
  id: string;
  image: string;
  name: string;
  model: string;
  type: string;
  createdDate: Date;
  updatedDate: Date;
  store: string;
  description: string;
}

export class MaterialEdit {
  id: string;
  image: string;
  name: string;
  model: string;
  type: string;
  store: string;
  description: string;
}